fix: correct relative import paths in API routes after directory restructure
When API routes were moved to index.ts subdirectories for trailingSlash
support, the relative imports became one level too shallow. Fixed all
four affected routes.
